pinning black dependency and adding lock file
Black only has pre-releases and pipenv allows either all pre-releases or
no pre-releases with the `--pre` switch.  To work around the issue we
are pinning the version of black in the Pipfile and not using --pre.

From now on we will need to update the versions of dependencies and
write a new .lock file.