Test an operation with bypassed authentication
Added a new API specification that is used only in tests. This
specification contains an operation that does not require
authentication. It is used in the new tests that verify that
authentication is not required for marked views.

This won’t be needed once there will be the unauthenticated
health check endpoint.