Explicitly mark authenticated views
To have more control on which operations require the identity header,
the authenticated view function must now be marked as such. All not
marked views are considered unauthorized, they are available without
the identity header and the identity object is not retrievable within
them. Introduced a requires_identity decorator that enables the identity
header validation and access for a given view.

There are no functional test, because the current implemenation does
not allow to add custom API to be used by the tests.