Count login failures
Add a new Prometheus Counter for login failures. Increment the counter
whenever an authentication fails: either if the header is undecodable
or invalid.

This doesn’t cover the case when the header is completely
missing. Such case is caught by Connexion and doesn’t get to the
authentication module at all.