Remove an overly complex test case
The TestApiRequestCountTestCase was too complicated given that it only
checks whether a specific Counter gets incremented. These would make
sense if the view functions were already thoroughly tested, but not if
these tests are the only onec there. The decorated is tested and thus
trusted, so there is no need to check its internals.