Add the API request count metric (#70)
* Add the API request count metric

Added a new inventory_request_count metric that counts the total
amount of API requests. It doesn’t include management (health and
metrics) requests. It increments on succeeded requests as well as on
failures.

* Change the decorator semantic

Changed the decorator from a generic “increase a given Counter” to
more semantic “this is an API request”. This makes the code simpler
and more descriptive. At the same time it opens possibilities to
extending it with more common functionality like time measurement and
request logging.

* Remove an overly complex test case

The TestApiRequestCountTestCase was too complicated given that it only
checks whether a specific Counter gets incremented. These would make
sense if the view functions were already thoroughly tested, but not if
these tests are the only onec there. The decorated is tested and thus
trusted, so there is no need to check its internals.

* PEP-8 fixes

* Removed unused imports.
* Put standard library imports above local imports.
* Removed whitespace from blank lines.