Use marshmallow for validation (#160)
This change moves the minLength and maxLength settings out of the openapi doc. This should allow the inventory service to fail just a single host instead of an entire bulk insert in the case that a single host has an invalid input string, invalid uuid, invalid ip address, invalid mac address.