Add pagination links to query outputs
Added pagination links (first, next, previous, last) to query outputs:
hosts and system profiles. The links are absolute paths, but not
URLs. The next and previous links can be missing if the page doesn’t
exist. The path and its query are preserved, only the _page_ and
_per_page_ parameters are filled. If not provided, parameters are
filled with their default values.