Extract (de)serialization methods
Moved all the Host, CanonicalFacts and Facts (de)serialization methods
from the classes to bare functions. The CanonicalFacts and Facts were
mere containers for static methods. The Host is a database model, which
should not be concerned about serialization.

The new bare methods can be easily extracted to a separate module. They
have unified naming including the argument names.