DRY the UUID generation
There are already functions with expressive names that convert the UUIDs
to strings either with or without hyphens. They were however used only
in a single test. Extracted those to a generic TestCase mixin and used
them everywhere where applicable.

The name is quite long in comparison to plain str(u) or u.hex, but it
clearly describes what is going on when used.