Fix trailing commas
Added trailing commas to the last item of multiline collections, if the
ending parenthesis is not of the same line. Removed the comma if the
parenthesis is there. This follows Black’s standard on where to put it.
Neither moved the parentheses/brackets in any way, nor changed the line
count of the expressions.