Modify the Host.update() method so that updating the system profile is optional. (#378)
Modify the Host.update() method so that updating the system profile is
optional.  This is a hack but its required at the moment.  The MQ
service needs to update the system profile when updating a host.  The
REST web service doesn't need to update the system profile when updating
a host.