Refactor Host constructor and tests
The logic of creating a Host model from JSON dict has been scattered
between the Host DB model constructor and the deserialization method.
Moved everything to the deserializer without changing any logic.

Refactored the (de)serialization tests to not work around the bugs and
quirks of the Host model constructor:

* Invalid default values for account and display_name
* Inability to provide all fields as keyword arguments

Added tests for the ability to fill an empty string into ansible_host.
This is no longer done by a common trusted method.