Replace sleep with flush
The correct way to wait for all pending Kafka messages to be produced is
to use the blocking method KafkaProducer.flush. Used this method instead
of current time.sleep, which is unreliable and may result in message
loss.

Furthermore, the time measurement doesn’t produce a right value. Without
the flush, it only measures how long does it take to add the messages to
buffer.