Fix host deletion JSON example formatting
Fixed the host deletion JSON example in the README, so it’s better
rendered by GitHub. Quoting the missing values results in them not being
painted red and it also adds the information that they are strings. Also
changed whitespace not to use hanging indentation and to start right at
the beginning of the line.