Mock emit_event in the race condition test
Currently, the test_delete_when_host_is_deleted test tries to use the
task subsystem, which is not initialized in tests and thus fails. Still,
we don’t want to emit any real events in the tests. Mocked out the
emit_event method, so the tests don’t fail and the subsystem isolated.