Merge master into split_inventory_service (#488)
* pin pipeline lib to previous stable version 1.3
recently there was some change to Jenkins shared pipeline lib a.k.a insights-pipeline that is causing some errors on inventory-smoke job. Temporally pinning this to the previous version till they get it stable again.
more info: https://github.com/RedHatInsights/insights-pipeline-lib/pull/40

* Move MockEmitEvent to class level

Having a class in a function is not necessary here. It just causes
Python to redeclare it every time the function is run instead of on the
class parse.

* Pass unformatted timestamp

DRY timestamp formatting. Pass it unformatted and call the isoformat
method right at the place where the other formatting happens.

* Check the host presence on the first GET

A GET request is currently done before deleting records. This request is
rather useless though, because 200 OK is returned even if no hosts are
found. Added an assert verifying that the host is actually found in the
database.

* Setting insights-pipeline-lib version due breaking changes

* Replace Mock.new with new_callable

unittest.Mock can use a callable (including a class constructor) to
create the mock. This is actually better than passing a manually created
class instance as we are sure that the instance is not shared anywhere.
Use as much from the Mock capabilities as possible.

* Pass host to the common test method

The shared _create_then_delete_host test method validates a host ID in
an emitted event, but it gets the delete URL determining the hosts to
delete from outside as an argument. Added a host argument to unify this.

* Split the delete tests class into three

The DeleteHostsTestCase class contains sets of rather unrelated tests.

- The race condition tests with a special interceptor class.
- The event tests with a special mock class.
- The invalid query tests that don’t require pre-created hosts.

Split into three separate classes.

* DOC: update README (#469)

* Dissect the delete event test

Split the common _create_then_delete_host method into smaller chunks.
This makes the tests more readable as the chunks are small and named. It
will also make it easier to completely get rid of the common method
without repeating too much logic.

* Added request_id test (#419)

Added the additional request_id test and query.all() in host.py.

* Extract datetime patching to setUp (#481)

Moved datetime return value patch to the shared _delete method and
timestamp to setUp. That gets us rid of passing around the timestamp and
digging it out of the mock.

* Extract host ID to the setUp method (#479)

Simplify tests by leveraging the setUp method. Extracted the host ID
there.

* Do not check request ID in every test (#474)

Check request ID only in request ID tests. This removes no request ID
test duplication and make the tests a little bit more atomic.

* Extract URL to the setUp method (#480)

Build the URL in the setUp method, so it’s not necessary to build it in
every test case and pass it around.

* Compose request ID header in tests (#476)

Move the request ID header composition from the DELETE method to the
request ID tests. Like this, the logic is at a more appropriate place.

* Remove the common test method (#475)

Remove the universal test method and call the stuff from its body
directly. Like this the tests are more straightforward.

* README - minor fixes (#472)

* Add ORDER BY id DESC to queries

To ensure unambiguous record ordering, oder by modified_on is not
enough. Order by primary key must be there so the same order is really
guaranteed. Added this ORDER BY clause to the host queries.

* Clean up the unittest imports

There were some unittest modules and submodules imports in test_api, but
they were used quite chaotically. Cleaned that up by always importing
and using the specific methods.

* DRY the invalid pagination parameter test

Wherever PaginationTestCase._base_paging_test is used, there is a query
that must accept valid pagination parameters and reject invalid ones.
Extracted the invalid pagination parameter test to PaginationTestCase
and used it in all places alongside _base_paging_test. This makes tests
more consistent.

* Fix two failing tests.