Add tags regression test
Added a tests for a case when hosts are queried by more tags in a single
namespace. Like that hosts with only the last tag in the query would be
matched even though they wouldn’t possess the other tags in the
namespace. This tests verifies that the bug is no longer present.