performance testing tools refactoring
* Refactoring payload building out of kafka_producer, and making it really easy to send multiple payloads (#401)

* Reformatted using black

* Replace sleep with flush

The correct way to wait for all pending Kafka messages to be produced is
to use the blocking method KafkaProducer.flush. Used this method instead
of current time.sleep, which is unreliable and may result in message
loss.

Furthermore, the time measurement doesn’t produce a right value. Without
the flush, it only measures how long does it take to add the messages to
buffer.

* Unify tools env with config

Made environment variable names and default values in the producer tool
with those in the Config class. This makes it possible to run the tool
with the same environment configuration.

* Remove unused build_chunk function

The build_chunk function is not used anywhere and it doesn’t do anything
useful. Removed entirely.

* Move executive code to main function

Modified the conusmer/producer scripts so they are more module-like. The
executive (side-effect) code is encapsulated in a main function. The
actual execution is under __name__ == "__main__" conditional.

In the payloads module, moved the RPM listing to a function, so it’s not
executed on import.

* Add TicToc to measure time

Use TicToc library to measure and print out elapsed time. It makes the
code more concise and readable.

* Extract random UUID function

Dried repeating str(uuid.uuid4()) code snippet to a single generate_uuid
function.

* Remove for abstraction from payloads

For-cycle is a notorious pattern and there is no need to abstract it to
the payloads module. Moved to the kafka_producer module, where it is
symmetric with the further MQ send iteration. It makes the code cleaner.

* Replace rpm call with a literal

Replaced a call to rpm with a literal list with ~300 packages. This
removed dependency on an external tool and thus adds support for
non-RHEL platforms.

* Reorder imports

Reordered imports so they are listed in a standard way. Split multiple
imports to single lines.

* Pass payload builder function

Remove the build_data function that decides what kind of payload is
built. Made build_payloads function to directly accept the builder
function as an argument. This allows to easier expansion, also prevents
dict mutation.

* Cleanup payloads variables and functions

* Removed unused variables, constants and functions.
* Ordered the objects so variables come before functions.
* Moved literal values and variables used only in one place to the
  respective place of usage.

* Refactor build_payloads method

Made the utils.payloads.build_payloads method more readable.

* Used the instance variant of str.encode method
* Explicitly stated the UTF-8 encoding
* Extracted a variable to distinguish a message from a payload

* Unify HTTP and MQ payload builidng

Use the same mechanics for building MQ and HTTP payloads. Every payload
type has its input method, both with the same interface and default
values.

* Fix pre-commit violations