Add host reaper script
Created a “reaper” script that removes all culled hosts. It is intended
to be run periodically to clean up the database. For every deleted host
a delete event is emitted.

Extracted the common host delete part that
handles the event emission and race condition handling.

The script uses SQLAlchemy and its models, but doesn’t use Flask.