add "captured_date" to system profile
This commit adds a "captured_date" field to the system profile. This
date records when the system profile was captured on the client
system.

For example, if a system upload was created on a Monday but uploaded
on Tuesday, this date would be Monday. PUP will populate this from the
system time in the uploaded tarball.