Add dateutil to Pipfile (#579)
* Add dateutil to Pipfile

python-dateutil is used to parse the ISO 8601 timestamps. This
library is already present as a deep dependency. There are some
explicit uses of dateutil in the tests. That could be potentially
a problem as n-th level dependency can disappear at any time
without notice. Fixed.

There is a bug making Connexion and Werzeug fail together. Upgrading
Connexion to 2.6.0 fixes the issue, but brings some breaking
changes. Some tests are now failing.

* Fix tests

- Change delete response code to 200 instead of newly default 204.
- Change expected missing field message to math the new format.

Co-authored-by: Ryan Dillin Felton <ryandillinfelton@gmail.com>