add "captured_date" to system profile (#578)
* add "captured_date" to system profile

This commit adds a "captured_date" field to the system profile. This
date records when the system profile was captured on the client
system.

For example, if a system upload was created on a Monday but uploaded
on Tuesday, this date would be Monday. PUP will populate this from the
system time in the uploaded tarball.

* Make the length limit a bit longer

ISO 8601 timestamp with a long timezone and six-digit second precision
is 32 characters long. PostgreSQL uses this precision. It may be a
good base stone. 2020-02-05T11:56:07.337557+00:00

Co-authored-by: Glutexo <driezasson@me.com>
Co-authored-by: Jozef Hartinger <jozefhartinger@gmail.com>