Allow query by tag without namespace
Made the namespace part of the tags query optional. This applies both
to the host query and to the tags query.

Added more tests for queries by tags with missing parts, especially for
the tags endpoint. In case of hosts query that uses PostgreSQL backend,
the current behavior is tested, even though it may be errorneous. Hosts
with a tag that has a value are matched even if the tags query contains
only the key and no value.