RHCLOUD-5321 RHCLOUD-3703 RHCLOUD-3819 Fix tags
Fixes
* https://projects.engineering.redhat.com/browse/RHCLOUD-5321
* https://projects.engineering.redhat.com/browse/RHCLOUD-3703
* https://projects.engineering.redhat.com/browse/RHCLOUD-3819


* Refactor Tag (de)composition and tests

Functions for converting Tag representations (string, nested) was a
mess. So was its test, although a bit less. Did some refactoring.

* Made the constructors static methods. Allows calling like
  Tag.from_*() instead of rather cryptic Tag().from_*().
* Refactored the insides so they don’t use too much confusing
  constructs and do leverage the power of Python and regular
  expressions.
* Split the huge test class into several atomic ones. Each tests only
  one method. Renamed the test methods not to repeat what the new test
  case class name says already.
* Moved asserts directly to the test methods, preferring readability
  over strict DRYness.

* Add more Tag tests

* Specifically encoded delimiters (/ and =) in all parts.
* Test maximum DECODED length of all parts.

* Allow non-encoded special chars in tags

Make the Tag string parsing regex match the rules in the Schema: Allow
anything except whitespace. Only delimiters need to be encoded.

* Allow special chars in tags query

Edited the regular expression so it accepts any undecoded character
except the delimiters and whitespace. That allows to use various
special characters like *, (, ) etc. without escaping. Added tests for
the host query tags parameter.

* Add tests for tags search

Added tests for the tags parameter of the tags query. Verifies that
special characters are accepted both unescaped and escaped and are
correctly passed to the Cross-Join query.

Also added tests verifying that it’s possible to add multiple tag
queries either as CSV or as multiple values.

* Allow query by tag without namespace

Made the namespace part of the tags query optional. This applies both
to the host query and to the tags query.

Added more tests for queries by tags with missing parts, especially for
the tags endpoint. In case of hosts query that uses PostgreSQL backend,
the current behavior is tested, even though it may be errorneous. Hosts
with a tag that has a value are matched even if the tags query contains
only the key and no value.

* Allow whitespace in tags

Removed whitespace restriction from tag instertion and search. It is
now possible to create host with whitespace in any tag part both by the
REST API and the MQ API. It is possible to search for tags with
namespace, it doesn’t even have to be escaped.

* Fix tests

Fix failing tests