RHCLOUD-4771 Make mandatory fields NOT NULL
* Make mandatory fields NOT NULL

Altered account, canonical_facts, created_on and modified_on to
NOT NULL. These fields are mandatory. Hosts with NULL or 'null' in
canonical_facts become {}. This is still invalid, but better.

RHCLOUD-4771

* Make data migration downgrade no-op

The data migration is fixing invalid records, not migrating data to a
new format. It doesn’t make sense thus to revert the fixed records back
to the invalid state. Made the downgrade do nothing.

* Merge data and DDL migrations

The DDL migration must happen immediately after fixing the data. Merged
the two migrations so they both happen in a same transaction. Extracted
the operations into functions so they are named and documented.

* Delete invalid records

There can still be records that don’t have a created/modified timestamp
or an account. There is no way to fix such hosts. Delete them.

* Log number of matched rows

Used standard inventory logging to introduce a new logging helper. Used
this helper to log the number of rows matched for UPDATE and DELETE.