Fix mock test assertion
Custom emit event mock class has been removed. As a result, emit_event
is a plain MagicMock and so is emit_event.events. Nothing is stored
there in the events attribute, but _len_ returns 0. That makes the test
pass, but the actual not emitting remains untested. Fixed by replacing
with assert_not_called.