RHCLOUD-4283 - Enabling stdout logs
* rebasing

* using logconfig as default log_config_file

* applying suggestions

* fixing cloudwatch logs

* cloudwatch working without global vars :tada:

* Stdout logging (#673)

* First working YAML config

Converted the fileConfig configuration file to a dictConfig compatible
YAML. This allows to remove KeyError to FileNotFoundError conversion
as dictConfig behaves more properly.

* Configure logging filter in config file

For the contextual logging filter to work with inventory.* child
loggers, it must be attached to handlers. Added the filter to the
configuration file, attaching it to the Logstash handler.

This still doesn’t enable the filter for the CloudWatch handler.

* fixup! Configure logging filter in config file

* Instantiate CloudWatch handler in config

Co-authored-by: Štěpán Tomsa <stomsa@redhat.com>

* remove cw handler if creds not present

* freaking lint :man_facepalming:

* reverting

* fix

* deleting logconfig.ini

* addressing suggestions

* Remove redundant logging

Removed CloudWatch configuration status logging. It doesn’t work
correctly and moreover duplicates existing print statements.

* Remove config file variable from Makefile

Value of INVENTORY_LOGGING_CONFIG_FILE in the Makefile only fills in
the default value. That is redundant and thus removed.

* update README

Co-authored-by: Shaun Dunning <shdunning@users.noreply.github.com>
Co-authored-by: Glutexo <stomsa@redhat.com>
Co-authored-by: Jozef Hartinger <jharting@redhat.com>