Set More Message Headers (RHCLOUD-5964) (#676)
* adding headers

* adding insights_id header

* fixing and updating tests

* revert

* adding insights_id headers + test updates

* removing unnecessary changes

* Don’t extract values from serialized host

Serialized host is an opaque value that should not be used for data
extraction. If any data in the serialized representation is required,
better practice is to pass this along, even though it looks like
unnecessary duplication.

* Do not mutate data passed to Event Producer

Event producer was mutating a passed dict argument to coalesce None
values to empty strings. Removed the mutation and made the coalesce
more generic.

* Do not extract Host ID from serialized Host

Just as with the Insights ID, extracting the Host ID from a serialized
payload is not a good practiced. Added Host ID to the places where
Insights ID has been added to avoid this practice. This makes the
serialized payloads opaque.

* Replace wordy conditionals with dict.get

`dict_["key"] if "key" in dict_ else None` can be replaced by
dict_.get("key"). Did this for the new retrievals of the Insights ID.

* Fix PATCH event emission

host.insights_id is invalid. host.canonical_facts["insights_id"] is correct.

* Add event test for Host update wo Insights ID

Message headers now contain Insights ID. Add a test verifying that
events for Hosts without one can still be created. Extract the Host
creation in PreCreatedHostsBaseTestCase, so more hosts with specific
data can be created in the test themselves.

* Add event test for Host delete wo Insights ID

Message headers now contain Insights ID. Add a test verifying that
events for Hosts without one can still be created.

* Fix MQ service tests

Added the Host ID argument to add_host mocks. Fixed (un)serialized
UUIDs in the tests.

* Fix failing facts tests

In the facts tests, the Culling test case was causing name conflicts
because of a complex inheritance chain. Untangled that and fixed the
failing tests.

Co-authored-by: Jozef Hartinger <jozefhartinger@gmail.com>
Co-authored-by: Štěpán Tomsa <stomsa@redhat.com>