Fix patch tests
In the patch tests, actual timestamps of a Host are compared to those
in the event. The expected ones were taken from a Host model instance
after it’s saved to the database. In that case, SQLAlchemy converts
the timestamp to local timezone, which then mismatches the timezone in
the produced event.Fixed that by converting the expected values back to UTC.