Graceful Shutdown Handler for Reaper - RHCLOUD-5358 (#690)
* Graceful shutdown for reaper + test

* modifying gitignore to allow files in from lib directory

* lint :expressionless:

* Graceful shutdown reaper (#691)

* Make host delete interrupt agnostic

Instead of coupling lib.host_delete to the shutdown handler
implementation, used a plain function to determine the interruption.
This increases concern separation.

* DRY signal handler registration

Moved signal handler registration to the handlers module. Code using
the handler needs not to know about the signals at all, it just
registers.

* Decouple queue from signal handler

Just as with the host delete, decoupled the queue event loop from the
signal event handling. Moved to the initialization module, passing in
only the function determining whether the event loop should continue.

* Fix tests

Moving the interrupt to the end of the loop results in a need for
different side effects in the tests. Fixed those and threw in a
little refactoring.

* Use atexit for shutdown handlers

To avoid shutdown handlers not being run in case of an exception, used
Python native atexit function to register those handlers. Extracted
the logic into lib.handlers.

Co-authored-by: Štěpán Tomsa <stomsa@redhat.com>

* fixing tests

* Consolidate shutdown handler mocks

Use return value in tests that don’t test the shutdown handler.

Co-authored-by: Štěpán Tomsa <stomsa@redhat.com>