Add a script that replaces IMAGE_TAG in deploy.yml (#693)
* First try

* Use regular expressions

* Name indexes

* Chmod deploy.py as executable

* Add deploy.py description to README

* Replace fixtures with helper functions

Removed fixtures that only introduce boiler plate code. Replaced them
with plain functions.

* Do not shadow variable names

args was used in the module and in the function scopes. Renamed.

* Rename test

Changed a test name to better describe what it does. If formatting
changes, it is not broken, even though it is undesirable.

* Fix the deploy command in readme

Forgot to include PROMO_CODE.