Test not using nested schema
It is not easily possible to patch a Marshmallow schema field. It’s
because they are defined on class declaration and not its
instantiation. Because of that added a crude test verifying that
a correct Schema is used for deserialization and that this host schema
does not directly use a nested Marshmallow schema for the system
profile.