RHCLOUD-8614 Use externalized schema for system profile validation (#700)
* Use system profile schema for validation

* Synchronize System Profile schema

With the only exception of /DiskDevice/options that does not use
recursive reference. It is not supported by Prance and can be
validated by Marshmallow.

* fixup! Synchronize System Profile schema

* Add type definitions

For type coercion to work correctly, objects need to have explicitly
declared types. Added type declarations to all objects in the schema.

* Check for empty keys in disk device options

JSON Schema version 4 can’t be used to validate object keys. Added a
manual validation for non-empty keys in disk_devices/options.

* fixup! Use system profile schema for validation

* Fix tests

Values strings lengths are now verified on the OpenAPI specification
level. This results in requests violating these constrains being
rejected before even reaching deserialization. They result thus in
immediate 400 Bad Request response instead of original 207
Multi-Status.

System Profile Disk Device Options are still validated manually. Empty
keys in this object don’t result in 400 Bad Request yet.

* Test constrains and type coercion

* fixup! Use system profile schema for validation

* Test not using nested schema

It is not easily possible to patch a Marshmallow schema field. It’s
because they are defined on class declaration and not its
instantiation. Because of that added a crude test verifying that
a correct Schema is used for deserialization and that this host schema
does not directly use a nested Marshmallow schema for the system
profile.

* Test use of externalized schema

Added a test that verifies that the externalized system profile schema
is used for validation. It patches the path to the specification and
presents a mock one instead that has more constrains. Then verifies
that the validation fails.

* Compose system profile schema on initialization

It is inefficient to compose big static dictionaries every time a
schema is used for validation. Moved the composition to the
constructor.

* Cache system profile schema

Used a class attribute to cache the System Profile schema. It is
loaded only on the first use. Required a test fix.

* Move schema reset to finally

If error occurs in the externalized System Profile test, the
MqHostSchema class will retain its modified schema. Added a
finally clause to ensure the class is reset even on error.

* Fix schema loading

* Remove undefined keys

Added object key filtering to the system profile processing. It
emulates Marshmallow behavior of removing keys that are not explicitly
defined.

This violates JSON Schema semantics and is a temporary measure until
"additionalProperties" is used to make the validation stricter. Only
objects with explicitly defined properties are affected, arbitrary
objects are kept unmodified. "additionalProperties" and "required"
directives are ignored.

* Revert unwanted payloads change

* Improve performance by object mutation

Replaced many uses of deepcopy by only single one. The functions used
in pre_load now mutate the passed dictionary. This is more efficient,
although less pure.

* Remove string and conditional repetition

Refactored the key filtering code to be more pythonic.

- Types are used instead of conditionals.
- Native exceptions are caught instead of preventing them.

Also removed string (schema dict keys and type values) repetition.
Introduced a mapping dictionary to list supported schema types.

* Fix unit tests

Updated key filtering unit tests, so they reflect new syntax. Fixed an
incorrect tests and added one more, so the test suite is more correct
and complete.

* Remove classes and exceptions

Use conditionals instead of exceptions to control the flow. Use plain
functions instead of types. This should increase performance.

* Add sap_sids constrains from schemas

Synchronized the new System Profile field from inventory-schemas.

* Flatten nested functions

Replaced the nested function paradigm with conditionals. This
made it possible to remove the void filter.

* Introduce Schema namedtuple

Created a namedtuple to represent the schema properties we care about.

* Put everything into a class

Create a class that serves as a namespace to avoid creating a module.

* Use an Enum for schema types

Replaced magic strings describing object type with an Enum. All
strings are now collected in a single place and only as lists.

* Fix unit tests

Unit tests were not adjusted to the new syntax. Fixed and tests are
now running correctly.

* Fix filtering in array items

Array branch contained a wrong function call. This was causing objects
nested in array items not to be filtered. Fixed.

* Simplify complex pieces

- Simplified None value filling.namedtuple.__new__.__defaults__ is no
  longer used to fall back to a default value.
- Removed namedtuple field property override. Access via self._asdict
  is not necessary anymore.
- Simplified known type check: Enum.__members__ is an OrderedDict. It
  is possibly use it with .get.

Co-authored-by: Jozef Hartinger <jozefhartinger@gmail.com>