Fix MQ System Profile tests
Message Queue API tests verifying System Profile normalization used
the generated event to verify the stored data. That was not correct,
because the outgoing event is subject to marshalling that does its
own normalization. That caused the unknown keys to be filtered even
though they were stored in the database and thus not testing our
custom filtering code. Fixed that.