RHCLOUD-7522 Fix system profile fields filtering (#715)
* Revert "Revert "RHCLOUD-8614 Use externalized schema for system profile validation (#700)""

This reverts commit 25c640f1

* Do not filter nested fields if not traversable

Key filtering failed if a value is described as array/object, but is
sent as a scalar. Fixed this by adding a type check for the actual
value.

* Fix yum_repos/base_url property name

Fixed a typo in the yum_repos property name. Inserted a missing
underscore.

* Extract schema loading to a util

Extracted System Profile schema loading to a utility function to be
reused. This is going to be utilized in the fake system profile tests.

* Extract schema mocking to a util

Extracted the System Profile schema file mocking to a utility function.
Added a unit test verifying the usage of the externalized schema.

* Fix key filtering

Object key filtering worked only for the root object, because JSON
schema references were not resolved. Used jsonschema.RefResolver to
dereference. That makes it possible to filter nested objects too.

* Refactor tests so they are more exhaustive

* Make System Profile normalizer a real class.

* Make System Profile normalizer a real class.

* Filter keys after validation

Moved the System Profile key filtering to @post_load. Like this it
corresponds more to the Marshmallow behavior. Also it allows to use
the System Profile schema to use additionalProperties for validation,
even though this is not used right now.

* Fix MQ System Profile tests

Message Queue API tests verifying System Profile normalization used
the generated event to verify the stored data. That was not correct,
because the outgoing event is subject to marshalling that does its
own normalization. That caused the unknown keys to be filtered even
though they were stored in the database and thus not testing our
custom filtering code. Fixed that.

* Fix invalid pytest configuration

system_profile_fixtures is no longer there. Causes tests to fail.

* Add missing invalid system profile tests

Added tests for fields missing in the list of profiles with too long
strings. Also added some tests for SAP IDs that violate its regex
pattern.

* Make base_url baseurl again

* Rename baseurl to base_url

yum_repos/baseurl was never saved. On the other hand, base_url was,
even though it was not defined in the specification. As a part of
putting the Marshallow constraints to the JSON Schema specification,
unified the field name to base_url.

* Add tests for base_url

Added more test to verify that the yum_repos/base_url field is handled
properly:

- The format: uri is not used, invalid urls pass.
- Length constraints are applied by Connexion.
- baseurl is neither validated nor saved by HTTP.
- baseurl is neither validated nor saved by MQ.

Co-authored-by: Tahmid Efaz <tefaz@redhat.com>